requirements.txt
cloudinary==1.8.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.2
django-widget-tweaks==1.4.1
gunicorn==19.7.1
mock==2.0.0
pbr==3.0.1
psycopg2==2.7.3.2
python-decouple==3.0
pytz==2017.2
six==1.10.0
static3==0.7.0
urllib3==1.21.1
virtualenv==15.1.0
